Give proper names for Lobby and Game DTOs
