+++ src/main/java/ch/uzh/ifi/seal/soprafs20/controller/GameController.java
@@ -4,10 +4,8 @@ package ch.uzh.ifi.seal.soprafs20.controller;
 import ch.uzh.ifi.seal.soprafs20.board.Board;
 import ch.uzh.ifi.seal.soprafs20.cards.Card;
 import ch.uzh.ifi.seal.soprafs20.field.Field;
-import ch.uzh.ifi.seal.soprafs20.game.Game;
 import ch.uzh.ifi.seal.soprafs20.rest.dto.ExchangePostDTO;
-import ch.uzh.ifi.seal.soprafs20.rest.dto.GameGetDTO;
-import ch.uzh.ifi.seal.soprafs20.rest.dto.GameGetDTO2;
+import ch.uzh.ifi.seal.soprafs20.rest.dto.LobbyGetDTO;
 import ch.uzh.ifi.seal.soprafs20.rest.dto.MovePostDTO;
 import ch.uzh.ifi.seal.soprafs20.rest.mapper.DTOMapper;
 import ch.uzh.ifi.seal.soprafs20.service.GameService;
@@ -42,8 +40,8 @@ public class GameController {
     @GetMapping("/game/{id}")
     @ResponseStatus(HttpStatus.OK)
     @ResponseBody
-    public GameGetDTO getGame(@PathVariable Long id) {
-        GameGetDTO gameNoDeck = DTOMapper.INSTANCE.convertEntityToGameGetDTO(gameService.getGameById(id));
+    public LobbyGetDTO getGame(@PathVariable Long id) {
+        LobbyGetDTO gameNoDeck = DTOMapper.INSTANCE.convertEntityToGameGetDTO(gameService.getGameById(id));
         return gameNoDeck;
     }
     //get current game with id
@@ -56,12 +54,12 @@ public class GameController {
     @ResponseStatus(HttpStatus.CREATED)
     @ResponseBody
 
-    public GameGetDTO exchangeCard(@RequestBody ExchangePostDTO exchangePostDTO,
-                             @RequestHeader("X-Token") String token, @PathVariable String id) {
+    public LobbyGetDTO exchangeCard(@RequestBody ExchangePostDTO exchangePostDTO,
+                                    @RequestHeader("X-Token") String token, @PathVariable String id) {
         User userExchangingCard = gameService.getUserByToken(token);
         long gameId = exchangePostDTO.getId();
         Card cardToExchange = exchangePostDTO.getCard();
-        GameGetDTO updatedGame =
+        LobbyGetDTO updatedGame =
                 DTOMapper.INSTANCE.convertEntityToGameGetDTO(gameService.letPlayersChangeCard(gameId, userExchangingCard.getId(),cardToExchange));
         return updatedGame;
     }
