+++ src/main/java/ch/uzh/ifi/seal/soprafs20/controller/LobbyController.java
@@ -1,9 +1,8 @@
 package ch.uzh.ifi.seal.soprafs20.controller;
 
 
-import ch.uzh.ifi.seal.soprafs20.rest.dto.GameGetDTO;
+import ch.uzh.ifi.seal.soprafs20.rest.dto.LobbyGetDTO;
 import ch.uzh.ifi.seal.soprafs20.user.User;
-import ch.uzh.ifi.seal.soprafs20.game.Game;
 import ch.uzh.ifi.seal.soprafs20.rest.dto.LobbyPostCreateDTO;
 import ch.uzh.ifi.seal.soprafs20.rest.dto.UserPostDTO;
 import ch.uzh.ifi.seal.soprafs20.service.GameService;
@@ -23,9 +22,9 @@ public class LobbyController {
     @GetMapping("/lobbies")
     @ResponseStatus(HttpStatus.OK)
     @ResponseBody
-    public List<GameGetDTO> getAllLobbies(@RequestHeader("X-Token") String token){
+    public List<LobbyGetDTO> getAllLobbies(@RequestHeader("X-Token") String token){
         if(gameService.checkIfUserExists(token)) {
-            List<GameGetDTO> allGames = gameService.getAllLobbies();
+            List<LobbyGetDTO> allGames = gameService.getAllLobbies();
             return allGames;
         }
         else{
@@ -36,11 +35,11 @@ public class LobbyController {
     @PostMapping("/lobbies")
     @ResponseStatus(HttpStatus.CREATED)
     @ResponseBody
-    public GameGetDTO createLobby(@RequestBody LobbyPostCreateDTO lobbyPostCreateDTO,
-                            @RequestHeader("X-Token") String token){
+    public LobbyGetDTO createLobby(@RequestBody LobbyPostCreateDTO lobbyPostCreateDTO,
+                                   @RequestHeader("X-Token") String token){
         String lobbyName = lobbyPostCreateDTO.getName();
         User userCreatingLobby = gameService.getUserByToken(token);
-        GameGetDTO createdGame = gameService.createLobby(userCreatingLobby, lobbyName);
+        LobbyGetDTO createdGame = gameService.createLobby(userCreatingLobby, lobbyName);
         return createdGame;
     }
 
@@ -48,7 +47,7 @@ public class LobbyController {
     @GetMapping("/lobby/{id}")
     @ResponseStatus(HttpStatus.OK)
     @ResponseBody
-    public GameGetDTO getLobbyById(@PathVariable Long id){
+    public LobbyGetDTO getLobbyById(@PathVariable Long id){
         return gameService.getLobbyById(id);
     }
