+++ src/main/java/ch/uzh/ifi/seal/soprafs20/service/GameService.java
@@ -3,7 +3,7 @@ package ch.uzh.ifi.seal.soprafs20.service;
 
 import ch.uzh.ifi.seal.soprafs20.cards.Value;
 import ch.uzh.ifi.seal.soprafs20.game.GameState;
-import ch.uzh.ifi.seal.soprafs20.rest.dto.GameGetDTO;
+import ch.uzh.ifi.seal.soprafs20.rest.dto.LobbyGetDTO;
 import ch.uzh.ifi.seal.soprafs20.rest.mapper.DTOMapper;
 import ch.uzh.ifi.seal.soprafs20.user.Colour;
 import ch.uzh.ifi.seal.soprafs20.user.Figure;
@@ -15,7 +15,6 @@ import ch.uzh.ifi.seal.soprafs20.field.Field;
 import ch.uzh.ifi.seal.soprafs20.game.Game;
 import ch.uzh.ifi.seal.soprafs20.repository.GameRepository;
 import ch.uzh.ifi.seal.soprafs20.rest.dto.MovePostDTO;
-import ch.uzh.ifi.seal.soprafs20.rest.dto.UserPostDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
@@ -292,7 +291,7 @@ public class GameService {
         return boardService.checkIfAllTargetFieldsOccupied(gameId, currentPlayer);
     }
 
-    public GameGetDTO getLobbyById(long id){
+    public LobbyGetDTO getLobbyById(long id){
         Game game = gameRepository.findById(id).orElse(null);
         return DTOMapper.INSTANCE.convertEntityToGameGetDTO(game);
     }
@@ -316,11 +315,11 @@ public class GameService {
         return userService.getUserByToken(tokenOfUser) != null;
     }
 
-    public List<GameGetDTO> getAllLobbies(){
+    public List<LobbyGetDTO> getAllLobbies(){
 
 
         List<Game> allGames = gameRepository.findAll();
-        List<GameGetDTO> games = new ArrayList<>();
+        List<LobbyGetDTO> games = new ArrayList<>();
         for(Game game : allGames){
             games.add(DTOMapper.INSTANCE.convertEntityToGameGetDTO(game));
         }
@@ -331,7 +330,7 @@ public class GameService {
         return userService.getUserByToken(token);
     }
 
-    public GameGetDTO createLobby(User userOwner, String gameName){
+    public LobbyGetDTO createLobby(User userOwner, String gameName){
         Game game = new Game(userOwner, gameName);
         gameRepository.saveAndFlush(game);
         return DTOMapper.INSTANCE.convertEntityToGameGetDTO(game);
