+++ src/test/java/ch/uzh/ifi/seal/soprafs20/controller/LobbyControllerTest.java
@@ -2,17 +2,14 @@ package ch.uzh.ifi.seal.soprafs20.controller;
 
 import ch.uzh.ifi.seal.soprafs20.game.Game;
 import ch.uzh.ifi.seal.soprafs20.game.WeatherState;
-import ch.uzh.ifi.seal.soprafs20.rest.dto.GameGetDTO;
+import ch.uzh.ifi.seal.soprafs20.rest.dto.LobbyGetDTO;
 import ch.uzh.ifi.seal.soprafs20.rest.dto.LobbyPostCreateDTO;
 import ch.uzh.ifi.seal.soprafs20.rest.mapper.DTOMapper;
 import ch.uzh.ifi.seal.soprafs20.service.GameService;
-import ch.uzh.ifi.seal.soprafs20.service.UserService;
-import ch.uzh.ifi.seal.soprafs20.user.Player;
 import ch.uzh.ifi.seal.soprafs20.user.User;
 import ch.uzh.ifi.seal.soprafs20.user.UserStatus;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.jayway.jsonpath.JsonPath;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,12 +19,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
-import org.springframework.util.Assert;
 import org.springframework.web.server.ResponseStatusException;
 
 import static org.hamcrest.Matchers.*;
 import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
@@ -58,9 +53,9 @@ public class LobbyControllerTest {
 
         Game lobby = new Game(user, "testGame");
 
-        GameGetDTO lobbyGetDTO = DTOMapper.INSTANCE.convertEntityToGameGetDTO(lobby);
+        LobbyGetDTO lobbyGetDTO = DTOMapper.INSTANCE.convertEntityToGameGetDTO(lobby);
 
-        List<GameGetDTO> allGames = Collections.singletonList(lobbyGetDTO);
+        List<LobbyGetDTO> allGames = Collections.singletonList(lobbyGetDTO);
 
         given(gameService.getAllLobbies()).willReturn(allGames);
         given(gameService.checkIfUserExists(Mockito.anyString())).willReturn(true);
@@ -88,11 +83,11 @@ public class LobbyControllerTest {
         LobbyPostCreateDTO lobbyPostCreateDTO = new LobbyPostCreateDTO();
         lobbyPostCreateDTO.setName("Allo");
 
-        GameGetDTO gameGetDTO = new GameGetDTO();
-        gameGetDTO.setName(lobbyPostCreateDTO.getName());
+        LobbyGetDTO lobbyGetDTO = new LobbyGetDTO();
+        lobbyGetDTO.setName(lobbyPostCreateDTO.getName());
 
         given(gameService.getUserByToken(Mockito.anyString())).willReturn(user);
-        given(gameService.createLobby(Mockito.any(), Mockito.anyString())).willReturn(gameGetDTO);
+        given(gameService.createLobby(Mockito.any(), Mockito.anyString())).willReturn(lobbyGetDTO);
 
         MockHttpServletRequestBuilder postRequest = post("/lobbies").contentType(MediaType.APPLICATION_JSON)
                 .content(asJsonString(lobbyPostCreateDTO))
@@ -114,7 +109,7 @@ public class LobbyControllerTest {
         lobby.setWeatherState(WeatherState.CASUAL);
         
 
-        GameGetDTO gameGetDTO = DTOMapper.INSTANCE.convertEntityToGameGetDTO(lobby);
+        LobbyGetDTO lobbyGetDTO = DTOMapper.INSTANCE.convertEntityToGameGetDTO(lobby);
 
         given(gameService.getLobbyById(1)).willReturn(DTOMapper.INSTANCE.convertEntityToGameGetDTO(lobby));
